Extracted logic for session completion, Added SessionHeader.tsx & SessionVideo.tsx components
